--- a/pkg/bus/bus.go
+++ b/pkg/bus/bus.go
@@ -17,6 +17,7 @@ type EventPublishFunc func(data.Event)
 type EventBus struct {
 	subscribers []EventReceiveFunc
 	rw          sync.RWMutex
+	wg          sync.WaitGroup
 }
 
 // Subscribe subscribe to bus
@@ -31,10 +32,13 @@ func (eb *EventBus) Publish(e data.Event) {
 	eb.rw.RLock()
 
 	for _, rf := range eb.subscribers {
-		go func(rf EventReceiveFunc) {
+		eb.wg.Add(1)
+		go func(rf EventReceiveFunc, wg *sync.WaitGroup) {
+			defer wg.Done()
 			rf(e)
-		}(rf)
+		}(rf, &eb.wg)
 	}
+	eb.wg.Wait()
 	eb.rw.RUnlock()
 }