This PR ensures events are processed in the order they arrived on the event bus by the application plugins.

I found out, that because the event bus runs the application plugin receiving functions inside a goroutine, sometimes events can be processed by the application plugins in a different order, than how they arrived on the bus. This is really bad especially for the Loki application plugin, because Loki requires the timestamps of logs to be increasing (last_log_timestamp <= new_log_timestamp). Because of this, sometimes a log sent to Loki would have an older timestamp, than the last log sent to Loki, which would cause "entry out of order" errors.

I wonder, if this behavior is ok for other storages we use (maybe we want to do a similar thing for metrics too). I'm not sure, if and how much, waiting for previous event to finish processing, could affect performance.