@vyzigold 2021-08-03
I was looking into where are we loosing logs yesterday. I tried sg-core with and without this patch. I got a lot more amqp_overruns in sg-bridge with this patch (i run sg-bridge with the --block flag), which looks to me like sg-core got noticably slower with this. I'd like to take some measurements today.