@pleimer 2021-08-04
Very interesting find.

For metrics, the timestamp created by collectd is used to store the metric into Prometheus, so whether or not a metric arrives slightly out of order does not affect the order it is stored in Prometheus. Side note, this could hinder Prometheus performance, but right now Prometheus is nowhere near the bottleneck in our system. In Elasticsearch, order of timestamps arriving doesn't matter either.

By adding a WG to force goroutines to execute in order, you are essentially re-creating the functionality of a channel. You may see slightly better performance if you utilize those instead. You may not, but it's probably worth testing.

I originally did performance testing for metrics with a bus based on buffered channels but reverted to the goroutine strategy because it was much more performant. It may be that we should change the event bus functionality only and leave the metric bus as is, but document for application plugin developers that messages may arrive out of order