@vyzigold 2021-08-06
So I tried to test the speed of this, but I don't think the results are really saying anything unless I test with ES directly (for Loki we need this PR and the application plugin feeds the data directly into a channel, so the speed should be similar). I wanted to test just the bus, so I wanted sg-core to do as little additional work as possible, so I chose the dummy-logs transport plugin and print application plugin to print into a file and just counted how many lines are inside the file, when all the lines were there, I stopped sg-core and measured the time. You can see my testing setup [here](https://github.com/vyzigold/sg-core/tree/bus-testing). I run the "stop" script in the background and then run `time ./sg-core`.
I run the tests multiple times, the results were:
without this PR:
./sg-core -config config.yaml  1033.43s user 68.91s system 367% cpu 5:00.32 total
./sg-core -config config.yaml  1034.51s user 66.30s system 383% cpu 4:46.84 total
./sg-core -config config.yaml  1044.94s user 68.87s system 367% cpu 5:03.06 total
./sg-core -config config.yaml  1037.47s user 63.69s system 377% cpu 4:51.92 total

with this PR:
./sg-core -config config.yaml  510.30s user 76.39s system 211% cpu 4:37.77 total
./sg-core -config config.yaml  496.79s user 73.57s system 212% cpu 4:28.09 total
./sg-core -config config.yaml  488.12s user 73.71s system 210% cpu 4:26.55 total

with channel:
./sg-core -config config.yaml  491.14s user 62.33s system 205% cpu 4:29.80 total
./sg-core -config config.yaml  501.29s user 63.78s system 205% cpu 4:35.23 total

With the PR the sg-core was actually faster probably because it took longer for the fs to synchronize the access to the file, when more goroutines tried to write to it. With channel the speed was comparable, to the speed with this PR, but channel doesn't allow reads from multiple application plugins.