@vyzigold 2021-08-17
I added a configuration option "blockEventBus" and when it's set to true, the blocking code is used. I thought about Paul's suggestion to block each application plugin separately, but I'm not sure what would the best way to do this be. I might take a look at this again when we revisit the bus architecture. And as I wrote yesterday we might not even need a blocking bus then. Example configuration to use this would be:
```
pluginDir: /tmp/plugins
logLevel: debug
blockEventBus: true
transports:
    - name: dummy-logs
      handlers:
      - name: logs
        config:
            timestampField: "@timestamp"
            messageField: "message"
            severityField: "severity"
            hostnameField: "host"
applications: 
  - name: print
```