@vyzigold 2021-08-23
@pleimer Thank you for the suggestion. I see the FIFO ordering problem you mentioned in your suggestion, but I think my solution should actually guarantee FIFO order, unless I missed something, each write to the event bus (in my solution) blocks the whole sg all the way to reading from socket and that should be FIFO as well (sg-bridge as far as I know is FIFO as well). I know this isn't ideal, but from my testing so far and from all of the issues with log ordering in the Loki community, I think we really should try to guarantee FIFO ordering if we want to write to Loki (If we don't guarantee FIFO ordering, there is probably no point in even trying to use Loki currently). I don't see a way to guarantee FIFO ordering unless we block everything all the way to the socket and at that point, it probably doesn't make sense to try to not block writes to separate applications (the faster applications would just don't have any data to process, because the socket plugin shouldn't give them anything).

I'd suggest to just use the blocking only for Loki and keep the logging sg separate (we need a different sg-bridge configuration for logs anyway infrawatch/smart-gateway-operator#91). The Loki community is currently working on dropping the ordering requirement, so in the future, if we want to consolidate smart gateways and Loki doesn't require the logs strictly ordered, we can revert this PR and just use the current implementation of the event bus