@pleimer 2021-08-24
Thanks the explanation @vyzigold . You're right there really isn't any other way to guarantee order and that none are dropped without blocking everything. Some may be dropped in sg-bridge after a while, but at  that point we would need to start discussing routing to multiple sgs.